+++ data/assignment.go
@@ -14,6 +14,12 @@ type AssignmentCfg struct {
 	MaxPerTeam int
 }
 
+type QuadraticAssignment struct {
+	Count    int
+	Distance [][]float64
+	Flow     [][]float64
+}
+
 // NewAssignment creates a new AssignmentCfg
 func NewAssignment(name string) *AssignmentCfg {
 	data, err := load(name)
@@ -45,5 +51,26 @@ func NewAssignment(name string) *AssignmentCfg {
 		Cost:       cost,
 		Teams:      teams,
 		MaxPerTeam: number.ParseInt(data["maxPerTeam"]),
+	}
+}
+
+// NewQuadraticAssignment creates a new Quadratic Assignment config
+func NewQuadraticAssignment(name string) *QuadraticAssignment {
+	data, err := load(name)
+	if err != nil {
+		return nil
+	}
+	distance := make([][]float64, 0)
+	for _, line := range parseList(data["distance"]) {
+		distance = append(distance, matrixRow(line, false))
+	}
+	flow := make([][]float64, 0)
+	for _, line := range parseList(data["flow"]) {
+		flow = append(flow, matrixRow(line, false))
+	}
+	return &QuadraticAssignment{
+		Count:    number.ParseInt(data["count"]),
+		Distance: distance,
+		Flow:     flow,
 	}
 }